@@ -1,19 +1,27 @@
 package com.baeldung.abstractclasses.filereaders;
 
 import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public abstract class BaseFileReader {
     
-    protected String filePath;
+    protected Path filePath;
     
-    protected BaseFileReader(String filePath) {
+    protected BaseFileReader(Path filePath) {
         this.filePath = filePath;
     }
     
-    public String getFilePath() {
+    public Path getFilePath() {
         return filePath;
     }
     
-    public abstract List<String> readFile() throws IOException;
+    public List<String> readFile() throws IOException {
+        return Files.lines(filePath)
+            .map(this::mapFileLine).collect(Collectors.toList());
+    }
+    
+    protected abstract String mapFileLine(String line);
 }
