@@ -1,21 +1,15 @@
 package com.baeldung.abstractclasses.filereaders;
 
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.List;
-import java.util.stream.Collectors;
+import java.nio.file.Path;
 
 public class LowercaseFileReader extends BaseFileReader {
 
-    public LowercaseFileReader(String filePath) {
+    public LowercaseFileReader(Path filePath) {
         super(filePath);
     }
-    
+
     @Override
-    public List<String> readFile() throws IOException {
-        return Files.lines(Paths.get(filePath))
-            .map(String::toLowerCase)
-            .collect(Collectors.toList());
-    }
+    public String mapFileLine(String line) {
+        return line.toLowerCase();
+    }   
 }
