@@ -4,7 +4,6 @@ import java.io.IOException;
 import java.io.PrintWriter;
 import java.net.ServerSocket;
 import java.net.Socket;
-import java.util.Date;
 
 import javax.net.ServerSocketFactory;
 import javax.net.ssl.SSLServerSocket;
@@ -13,6 +12,7 @@ import javax.net.ssl.SSLServerSocketFactory;
 public class SimpleServer {
     static void startServer(int port) throws IOException {
         ServerSocketFactory factory = SSLServerSocketFactory.getDefault();
+
         try (ServerSocket listener = factory.createServerSocket(port)) {
             ((SSLServerSocket) listener).setNeedClientAuth(true);
             ((SSLServerSocket) listener).setEnabledCipherSuites(
@@ -22,13 +22,13 @@ public class SimpleServer {
             while (true) {
                 try (Socket socket = listener.accept()) {
                     PrintWriter out = new PrintWriter(socket.getOutputStream(), true);
-                    out.println(new Date().toString());
+                    out.println("Hello World!");
                 }
             }
         }
     }
 
     public static void main(String[] args) throws IOException {
-        startServer(1234);
+        startServer(8443);
     }
 }
