@@ -10,7 +10,7 @@ public class DeptEmployee {
 
     private String employeeNumber;
 
-    private String designation;
+    private String title;
 
     private String name;
 
@@ -23,11 +23,11 @@ public class DeptEmployee {
         this.department = department;
     }
     
-    public DeptEmployee(String name, String employeeNumber, String designation, Department department) {
+    public DeptEmployee(String name, String employeeNumber, String title, Department department) {
         super();
         this.name = name;
         this.employeeNumber = employeeNumber;
-        this.designation = designation;
+        this.title = title;
         this.department = department;
     }
 
@@ -63,11 +63,11 @@ public class DeptEmployee {
         this.department = department;
     }
 
-    public String getDesignation() {
-        return designation;
+    public String getTitle() {
+        return title;
     }
 
-    public void setDesignation(String designation) {
-        this.designation = designation;
+    public void setTitle(String title) {
+        this.title = title;
     }
 }
