@@ -30,7 +30,7 @@ public class ContactInfoValidator implements ConstraintValidator<ContactInfo, St
         if (StringUtils.isEmptyOrWhitespace(expressionType)) {
             LOG.error("Contact info type missing!");
         } else {
-            pattern = expressionRepository.findOne(expressionType).map(ContactInfoExpression::getPattern).orElse("");
+            pattern = expressionRepository.findById(expressionType).map(ContactInfoExpression::getPattern).orElse("");
         }
     }
 
