@@ -7,5 +7,5 @@ import org.springframework.data.repository.Repository;
 import com.baeldung.dynamicvalidation.model.ContactInfoExpression;
 
 public interface ContactInfoExpressionRepository extends Repository<ContactInfoExpression, String> {
-    Optional<ContactInfoExpression> findOne(String id);
+    Optional<ContactInfoExpression> findById(String id);
 }
