@@ -1,10 +1,10 @@
 package com.baeldung.rss;
 
+import javax.annotation.security.RolesAllowed;
+
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
-
-import javax.annotation.security.RolesAllowed;
 
 @SpringBootApplication
 @ComponentScan(basePackages = "com.baeldung.rss")
@@ -12,7 +12,6 @@ public class RssApp {
 
     @RolesAllowed("*")
     public static void main(String[] args) {
-        System.setProperty("security.basic.enabled", "false");
         SpringApplication.run(RssApp.class, args);
     }
 
