@@ -1,31 +1,34 @@
 package com.baeldung.displayallbeans;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.BDDAssertions.then;
+
+import java.net.URI;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.actuate.beans.BeansEndpoint.ContextBeans;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.RequestEntity;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.web.context.WebApplicationContext;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-
-import static org.assertj.core.api.BDDAssertions.then;
-import static org.hamcrest.CoreMatchers.hasItem;
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertTrue;
-
 @RunWith(SpringRunner.class)
 @SpringBootTest(classes = Application.class, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@TestPropertySource(properties = { "management.port=0", "endpoints.beans.id=springbeans", "endpoints.beans.sensitive=false" })
+@TestPropertySource(properties = { "management.port=0", "management.endpoints.web.exposure.include=*" })
 public class DisplayBeanIntegrationTest {
 
     @LocalServerPort
@@ -40,6 +43,8 @@ public class DisplayBeanIntegrationTest {
     @Autowired
     private WebApplicationContext context;
 
+    private static final String ACTUATOR_PATH = "/actuator";
+
     @Test
     public void givenRestTemplate_whenAccessServerUrl_thenHttpStatusOK() throws Exception {
         ResponseEntity<String> entity = this.testRestTemplate.getForEntity("http://localhost:" + this.port + "/displayallbeans", String.class);
@@ -49,22 +54,27 @@ public class DisplayBeanIntegrationTest {
 
     @Test
     public void givenRestTemplate_whenAccessEndpointUrl_thenHttpStatusOK() throws Exception {
-        @SuppressWarnings("rawtypes")
-        ResponseEntity<List> entity = this.testRestTemplate.getForEntity("http://localhost:" + this.mgt + "/springbeans", List.class);
+        ParameterizedTypeReference<Map<String, ContextBeans>> responseType = new ParameterizedTypeReference<Map<String, ContextBeans>>() {
+        };
+        RequestEntity<Void> requestEntity = RequestEntity.get(new URI("http://localhost:" + this.mgt + ACTUATOR_PATH + "/beans"))
+            .accept(MediaType.APPLICATION_JSON)
+            .build();
+        ResponseEntity<Map<String, ContextBeans>> entity = this.testRestTemplate.exchange(requestEntity, responseType);
 
         then(entity.getStatusCode()).isEqualTo(HttpStatus.OK);
     }
 
     @Test
     public void givenRestTemplate_whenAccessEndpointUrl_thenReturnsBeanNames() throws Exception {
-        @SuppressWarnings("rawtypes")
-        ResponseEntity<List> entity = this.testRestTemplate.getForEntity("http://localhost:" + this.mgt + "/springbeans", List.class);
+        RequestEntity<Void> requestEntity = RequestEntity.get(new URI("http://localhost:" + this.mgt + ACTUATOR_PATH + "/beans"))
+            .accept(MediaType.APPLICATION_JSON)
+            .build();
+        ResponseEntity<BeanActuatorResponse> entity = this.testRestTemplate.exchange(requestEntity, BeanActuatorResponse.class);
 
-        List<Map<String, Object>> allBeans = (List) ((Map) entity.getBody().get(0)).get("beans");
-        List<String> beanNamesList = allBeans.stream().map(x -> (String) x.get("bean")).collect(Collectors.toList());
+        Collection<String> beanNamesList = entity.getBody()
+            .getBeans();
 
-        assertThat(beanNamesList, hasItem("fooController"));
-        assertThat(beanNamesList, hasItem("fooService"));
+        assertThat(beanNamesList).contains("fooController", "fooService");
     }
 
     @Test
@@ -72,7 +82,20 @@ public class DisplayBeanIntegrationTest {
         String[] beanNames = context.getBeanDefinitionNames();
 
         List<String> beanNamesList = Arrays.asList(beanNames);
-        assertTrue(beanNamesList.contains("fooController"));
-        assertTrue(beanNamesList.contains("fooService"));
+        assertThat(beanNamesList).contains("fooController", "fooService");
+    }
+
+    private static class BeanActuatorResponse {
+        private Map<String, Map<String, Map<String, Map<String, Object>>>> contexts;
+
+        public Collection<String> getBeans() {
+            return this.contexts.get("application")
+                .get("beans")
+                .keySet();
+        }
+
+        public Map<String, Map<String, Map<String, Map<String, Object>>>> getContexts() {
+            return contexts;
+        }
     }
 }
